[blockwise quant] update scale to be f32

+++ src/xnnpack/microparams-init.h
@@ -131,7 +131,7 @@ XNN_INTERNAL void xnn_init_qs8_qc8w_bl_scale_fp32_params(
   size_t block_stride,
   // How much offset to subtract from packed_w pointer when moving from channels_tile to channels_subtile.
   size_t stride_offset,
-  const int32_t scale[XNN_MIN_ELEMENTS(1)],
+  const float scale[XNN_MIN_ELEMENTS(1)],
   void* packed_w);
